Update Dockerfile
This takes advantage of multi-stage build to both reduce the size of the
container as well as the number of files on the final container.

The build container used contains gradle and all the dependencies
whereas the final container only contains the .jar file it needs to run
the server.

One breaking change is that the configuration is no longer expected to
be within `/root/config`. For security best practices the application
should never run as `root` and therefore it will not have access to the
`/root/config` directory. The new user, `app`, will only have access to
`/app/config` and this is the new location where the configuration
should be mounted in.